Extract library-page routes into blueprints/library_routes.py (PR C of quickstart.py refactor)
12 routes + helpers move out of quickstart.py into a dedicated blueprint:

* /get_top_imdb_items/<library_name>
* /library_fragment/<library_id>
* /autosave_library/<library_id>
* /libraries_{tautulli,omdb,mdblist,anidb,radarr,sonarr,trakt,mal}_dependency_hint (x8)
* /copy_library_settings

Plus 5 supporting helpers: _configured_library_ids, _build_library_lists,
_legacy_playlist_library_names, _migrate_legacy_playlist_libraries_to_library_toggles,
_build_merged_libraries_hint_payload, _libraries_dependency_hint_response.

This PR builds on PR #1321 (workspace_status + dependency_reasons extraction).
The new blueprint imports the _libraries_data_*_dependency_reasons family
directly from modules.dependency_reasons -- no more qs_module indirection.

## DRY refactors landed alongside the move

The audit specifically called out two DRY opportunities; both are implemented here:

1. **7 nearly-identical dependency-hint routes**. The pre-refactor code had
   seven 9-line copy/paste blocks differing only in service label, URL, and
   resolver. These collapse into a (label, url-slug, resolver) registry table
   driven by bp.add_url_rule() with a closure factory. Each route still has a
   real Python function with the correct __name__, __doc__, and Flask
   endpoint name -- so error messages and URL resolution are unchanged.

2. **Inline _library_prefix() helper inside copy_library_settings**. Removed
   in favor of _library_prefix_from_key from modules.dependency_reasons.
   The module function is a strict superset of the inline one (it handles
   per-media-type markers like -movie-template_ / -show-overlay_ that the
   inline version missed), so this is also a small correctness improvement.

## Cross-module callees that stay in quickstart.py

Four library-validation helpers (_validate_library_collection_files,
_validate_library_metadata_files, _validate_library_overlay_files,
_validate_library_auto_sort_hubs), the file-entry normalizer
(_normalize_library_file_entries_payload), the cloner
(_clone_library_file_entries_for_target), and the selected-id resolver
(_selected_library_ids_from_libraries_data) are NOT yet extracted -- they
belong to PR D (library_file_entries module). The blueprint calls them via
a lazy  inside the two route bodies that need
them (autosave_library, copy_library_settings). Once PR D lands these
calls migrate to imports from modules.library_file_entries.

## Test compatibility

* All 791 tests pass; 267 directly exercise this code path.
* New fixture library_routes_module in tests/conftest.py for cases where a
  test needs to monkeypatch a helper that is *called from inside* a
  blueprint route (same pattern as workspace_status_module in PR A).
* Two tests in test_core_backend.py updated:
  - test_library_fragment_disables_auto_sort_hubs_without_plex_pass
  - test_copy_library_settings_mirrors_metadata_files
  Both moved their qs_module patches to library_routes_module patches.

## Size impact

* quickstart.py: 9,831 -> 9,227 lines (-604, -6.1%)
* blueprints/library_routes.py: 596 lines (new, under the 600 soft-limit)

Cumulative across PR A + PR C: 11,087 -> 9,227 (-1,860, -16.8%).
